Refactor MCP server structure for improved modularity and maintainability
- Split `server.py` into multiple modules: `utils.py`, `tools/`, `resources/`, and `prompts/` for better organization.
- Implemented utility functions for loading tips from JSON in `utils.py`.
- Created dedicated modules for time-related tools, greeting tools, and tips tools.
- Registered all components in the main server file, enhancing clarity and separation of concerns.
