+++ src/prompts/__init__.py
@@ -1,0 +1,1 @@
+# Prompts module for MCP server
