+++ src/resources/__init__.py
@@ -1,0 +1,1 @@
+# Resources module for MCP server
