+++ src/tools/__init__.py
@@ -1,0 +1,1 @@
+# Tools module for MCP server
